+++ changelog.md
@@ -4,6 +4,7 @@ changelog
 0.1.14 (development)
 --------------------
  * smarter rsync arguments
+ * now killing apps before restore on all platforms (needed when rsync is used for restore instead of cp)
  * started preferences: custom paths for backupfolder, logfile, rsync binary and busybox binary
  * detect changes made from batchactivity